+++ src/common/ExportSettingsModal/ExportSettingsModalContainer.js
@@ -1,20 +1,31 @@
-import React, { useCallback, useState } from 'react';
-import PropTypes from 'prop-types';
-import { useIntl } from 'react-intl';
 import moment from 'moment';
-
-import { stripesConnect } from '@folio/stripes/core';
-import { exportToCsv } from '@folio/stripes/components';
+import PropTypes from 'prop-types';
 import {
+  useCallback,
+  useMemo,
+  useState,
+} from 'react';
+import { useIntl } from 'react-intl';
+
+import { exportToCsv } from '@folio/stripes/components';
+import { stripesConnect } from '@folio/stripes/core';
+import { useCustomFields } from '@folio/stripes/smart-components';
+import {
+  CUSTOM_FIELDS_ORDERS_BACKEND_NAME,
   useShowCallout,
 } from '@folio/stripes-acq-components';
 
-import { getExportData, exportManifest } from './utils';
 import ExportSettingsModal from './ExportSettingsModal';
 import {
-  EXPORT_LINE_FIELDS,
-  EXPORT_ORDER_FIELDS,
-} from './constants';
+  exportManifest,
+  getExportData,
+  getExportLineFields,
+  getExportOrderFields,
+} from './utils';
+import {
+  ENTITY_TYPE_ORDER,
+  ENTITY_TYPE_PO_LINE,
+} from '../constants';
 
 const ExportSettingsModalContainer = ({
   onCancel,
@@ -22,6 +33,13 @@ const ExportSettingsModalContainer = ({
   fetchOrdersAndLines,
 }) => {
   const [isExporting, setIsExporting] = useState(false);
+  const [customFieldsPO, isLoadingPO] = useCustomFields(CUSTOM_FIELDS_ORDERS_BACKEND_NAME, ENTITY_TYPE_ORDER);
+  const [customFieldsPOL, isLoadingPOL] = useCustomFields(CUSTOM_FIELDS_ORDERS_BACKEND_NAME, ENTITY_TYPE_PO_LINE);
+  const customFields = useMemo(
+    () => [...(customFieldsPO || []), ...(customFieldsPOL || [])],
+    [customFieldsPO, customFieldsPOL],
+  );
+  const isLoadingCustomFields = isLoadingPO || isLoadingPOL;
   const showCallout = useShowCallout();
   const intl = useIntl();
 
@@ -32,7 +50,7 @@ const ExportSettingsModalContainer = ({
 
       const { lines, orders } = await fetchOrdersAndLines();
 
-      const exportData = await getExportData(mutator, lines, orders, intl);
+      const exportData = await getExportData(mutator, lines, orders, customFields, intl);
 
       setIsExporting(false);
 
@@ -41,7 +59,7 @@ const ExportSettingsModalContainer = ({
       const filename = `order-export-${moment().format('YYYY-MM-DD-hh:mm')}`;
 
       return exportToCsv(
-        [{ ...EXPORT_ORDER_FIELDS, ...EXPORT_LINE_FIELDS }, ...exportData],
+        [{ ...getExportOrderFields(customFields), ...getExportLineFields(customFields) }, ...exportData],
         {
           onlyFields: exportFields,
           header: false,
@@ -58,11 +76,13 @@ const ExportSettingsModalContainer = ({
     }
   },
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  [fetchOrdersAndLines, showCallout, onCancel]);
+  [customFields, fetchOrdersAndLines, showCallout, onCancel]);
 
   return (
     <ExportSettingsModal
+      customFields={customFields}
       isExporting={isExporting}
+      isLoading={isLoadingCustomFields}
       onExportCSV={onExportCSV}
       onCancel={onCancel}
     />