+++ src/common/ExportSettingsModal/utils/createExportReport.test.js
@@ -18,6 +18,70 @@ import {
   exportReport,
 } from '../../../../test/jest/fixtures';
 
+const customFieldsOrder = {
+  customFields: {
+    singleselect: 'opt_1',
+    textfield: 'abc',
+  },
+};
+
+const customFieldsOrderLine = {
+  customFields: {
+    multiselect: ['opt_1', 'opt_0', 'opt_3'],
+  },
+};
+
+const customFieldsResolved = {
+  singleselect: 'Value 2',
+  textfield: 'abc',
+  multiselect: 'A|B|opt_3',
+};
+
+const customFields = [
+  {
+    refId: 'singleselect',
+    name: 'Single select',
+    selectField: {
+      options: {
+        values: [
+          {
+            id: 'opt_1',
+            value: 'Value 2',
+          },
+          {
+            id: 'opt_0',
+            value: 'Value 1',
+          },
+        ],
+        sortingOrder: 'CUSTOM',
+      },
+    },
+  },
+  {
+    refId: 'textfield',
+    name: 'Text field',
+  },
+  {
+    refId: 'multiselect',
+    name: 'Multi select',
+    selectField: {
+      options: {
+        values: [
+          {
+            id: 'opt_1',
+            value: 'B',
+          },
+          {
+            id: 'opt_0',
+            value: 'A',
+          },
+        ],
+        sortingOrder: 'CUSTOM',
+      },
+    },
+  },
+];
+
 describe('createExportReport', () => {
   it('should return export report object', () => {
     const { result } = renderHook(() => useIntl());
@@ -27,6 +91,7 @@ describe('createExportReport', () => {
       intl,
       [orderLine],
       [order],
+      [], // customFields
       [vendor],
       [user],
       [acqUnit],
@@ -40,6 +105,20 @@ describe('createExportReport', () => {
       [{ id: orderLine.acquisitionMethod, value: 'Purchase' }],
       [{ id: vendor?.organizationTypes?.[0], name: 'Test type' }],
     )).toEqual(expect.objectContaining(exportReport));
+  });
+
+  it('should return export report object with custom fields', () => {
+    const { result } = renderHook(() => useIntl());
+    const intl = result.current;
+
+    const [exportRow] = createExportReport(
+      intl,
+      [{ ...orderLine, ...customFieldsOrderLine }],
+      [{ ...order, ...customFieldsOrder }],
+      customFields,
+    );
+
+    expect(exportRow.customFields).toEqual(customFieldsResolved);
   });
 
   it('should build rows with orders data even if there are no PO Lines in the order', () => {