+++ src/common/ExportSettingsModal/utils/index.js
@@ -1,3 +1,4 @@
 export * from './createExportReport';
 export * from './exportManifest';
 export * from './getExportData';
+export * from './exportFieldUtils';